优化：试题id转为雪花id

+++ Sourcecode/Song.Entities/LogForStudentQuestions.cs
@@ -24,7 +24,7 @@ namespace Song.Entities {
     		
     		protected DateTime _Lsq_LastTime;
     		
-    		protected Int32 _Qus_ID;
+    		protected Int64 _Qus_ID;
     		
     		protected Int32 _Lsq_Index;
     		
@@ -108,7 +108,7 @@ namespace Song.Entities {
     			}
     		}
     		
-    		public Int32 Qus_ID {
+    		public Int64 Qus_ID {
     			get {
     				return this._Qus_ID;
     			}
@@ -213,7 +213,7 @@ namespace Song.Entities {
     				this._Lsq_LastTime = reader.GetDateTime(_.Lsq_LastTime);
     			}
     			if ((false == reader.IsDBNull(_.Qus_ID))) {
-    				this._Qus_ID = reader.GetInt32(_.Qus_ID);
+    				this._Qus_ID = reader.GetInt64(_.Qus_ID);
     			}
     			if ((false == reader.IsDBNull(_.Lsq_Index))) {
     				this._Lsq_Index = reader.GetInt32(_.Lsq_Index);
@@ -285,7 +285,7 @@ namespace Song.Entities {
     			public static WeiSha.Data.Field Lsq_LastTime = new WeiSha.Data.Field<LogForStudentQuestions>("Lsq_LastTime");
     			
     			/// <summary>
-    			/// 字段名：Qus_ID - 数据类型：Int32
+    			/// 字段名：Qus_ID - 数据类型：Int64
     			/// </summary>
     			public static WeiSha.Data.Field Qus_ID = new WeiSha.Data.Field<LogForStudentQuestions>("Qus_ID");
